+++ projects/anviltech/wuf-ang-smart-table/src/lib/components/tbody/tbody.component.ts
@@ -3,7 +3,8 @@
  * Licensed under the MIT Open Source: https://opensource.org/licenses/MIT
  */
 
-import { Component, Input, Output, EventEmitter, ViewEncapsulation } from '@angular/core';
+import { Component, Input, Output, EventEmitter, ViewEncapsulation, OnChanges } from '@angular/core';
+import { TranslateService } from '@ngx-translate/core';
 
 import { Grid } from '../../services/grid';
 import { DataSource } from '../../data-source/data-source';
@@ -15,7 +16,7 @@ import { DataSource } from '../../data-source/data-source';
     templateUrl: './tbody.component.html',
     encapsulation: ViewEncapsulation.None
 })
-export class WufSmartTableTbodyComponent {
+export class WufSmartTableTbodyComponent implements OnChanges{
 
     @Input() grid: Grid;
     @Input() source: DataSource;
@@ -43,15 +44,24 @@ export class WufSmartTableTbodyComponent {
     isActionEdit: boolean;
     isActionDelete: boolean;
     noDataMessage: string;
+    columnCount: number;
+
+    constructor(public translate: TranslateService) {}
 
     ngOnChanges() {
+        this.mode = this.grid.getSetting('mode');
         this.isMultiSelectVisible = this.grid.isMultiSelectVisible();
         this.showActionColumnLeft = this.grid.showActionColumn('left');
-        this.mode = this.grid.getSetting('mode');
         this.showActionColumnRight = this.grid.showActionColumn('right');
-        this.isActionAdd = this.grid.getSetting('actions.add');
+        this.isActionAdd = this.grid.getSetting('actions.add.include');
         this.isActionEdit = this.grid.getSetting('actions.edit');
         this.isActionDelete = this.grid.getSetting('actions.delete');
         this.noDataMessage = this.grid.getSetting('noDataMessage');
+
+        // Determine column count
+        this.columnCount = this.grid.getColumns().length
+            + (this.isMultiSelectVisible ? 1 : 0)
+            + (this.showActionColumnLeft ? 1 : 0)
+            + (this.showActionColumnRight ? 1 : 0);
     }
 }